First refactor: change of containers file
