+++ client/containers/analysis-page/index.ts
@@ -1,5 +1,5 @@
-export * from "./header-section"
-export * from "./insights-section"
-export * from "./most-common-position-section"
-export * from "./recomendation-section"
-export * from "./overal-analysis-section"
+export * from "./header-section/analysis-header"
+export * from "./insights-section/insights-section"
+export * from "./most-common-position-section/most-common-position-section"
+export * from "./recomendation-section/recommendation-section"
+export * from "./overal-analysis-section/overal-analysis-section"
