+++ client/containers/analysis-page/insights-section/insights-section.tsx
@@ -1,4 +1,4 @@
-import { PayGapPieChart } from "@/components/PayGapChart/idex";
+import { PayGapPieChart } from "@/components/PayGapChart/PayGapPieChart";
 import { DataContext } from "@/services/provider";
 import { splitTextIntoTwoParts } from "@/services/splitTextIntoTwoParts";
 import React, { useContext } from "react";
